+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ﻿# Changelog
 ## Version 2.7.8
+* Added custom clutter, check out the tutorial at https://valheim-modding.github.io/Jotunn/tutorials/zones.html#adding-clutter
 * Improved mod compatibility window, the disconnect reason is shown inside the Jotunn window
 * Fixed mocking of some textures loaded by the game after vanilla prefabs are available